Add files via upload
The Car Rental Management System is a console-based C application that allows users to rent and return cars efficiently. It uses file handling to store and retrieve user and rental data securely. The system calculates the total cost based on the rental period and verifies user credentials during the return process. It demonstrates modular programming and structure-based data management in C.